+++ api/src/main/java/bio/terra/tanagra/app/controller/EntitiesApiController.java
@@ -24,9 +24,9 @@ import bio.terra.tanagra.proto.underlay.EnumHint;
 import bio.terra.tanagra.proto.underlay.EnumHintValue;
 import bio.terra.tanagra.proto.underlay.FilterableAttribute;
 import bio.terra.tanagra.proto.underlay.IntegerBoundsHint;
-import bio.terra.tanagra.service.underlay.EntityFiltersSchema;
-import bio.terra.tanagra.service.underlay.Underlay;
 import bio.terra.tanagra.service.underlay.UnderlayService;
+import bio.terra.tanagra.underlay.EntityFiltersSchema;
+import bio.terra.tanagra.underlay.Underlay;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
