+++ api/src/main/java/bio/terra/tanagra/app/controller/UnderlaysApiController.java
@@ -6,8 +6,8 @@ import bio.terra.tanagra.common.Paginator.Page;
 import bio.terra.tanagra.generated.controller.UnderlaysApi;
 import bio.terra.tanagra.generated.model.ApiListUnderlaysResponse;
 import bio.terra.tanagra.generated.model.ApiUnderlay;
-import bio.terra.tanagra.service.underlay.Underlay;
 import bio.terra.tanagra.service.underlay.UnderlayService;
+import bio.terra.tanagra.underlay.Underlay;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
