+++ api/src/main/java/bio/terra/tanagra/service/query/QueryService.java
@@ -9,8 +9,8 @@ import bio.terra.tanagra.service.search.SearchContext;
 import bio.terra.tanagra.service.search.SearchEngine;
 import bio.terra.tanagra.service.search.Selection;
 import bio.terra.tanagra.service.search.SqlVisitor;
-import bio.terra.tanagra.service.underlay.Underlay;
 import bio.terra.tanagra.service.underlay.UnderlayService;
+import bio.terra.tanagra.underlay.Underlay;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import java.util.Optional;
