+++ api/src/main/java/bio/terra/tanagra/service/query/api/ApiConversionService.java
@@ -9,8 +9,8 @@ import bio.terra.tanagra.service.query.EntityDataset;
 import bio.terra.tanagra.service.query.EntityFilter;
 import bio.terra.tanagra.service.search.EntityVariable;
 import bio.terra.tanagra.service.search.Filter;
-import bio.terra.tanagra.service.underlay.Underlay;
 import bio.terra.tanagra.service.underlay.UnderlayService;
+import bio.terra.tanagra.underlay.Underlay;
 import com.google.common.collect.ImmutableList;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
