+++ api/src/main/java/bio/terra/tanagra/service/search/SqlVisitor.java
@@ -6,13 +6,13 @@ import bio.terra.tanagra.service.search.Filter.ArrayFunction;
 import bio.terra.tanagra.service.search.Filter.BinaryFunction;
 import bio.terra.tanagra.service.search.Filter.RelationshipFilter;
 import bio.terra.tanagra.service.search.Selection.PrimaryKey;
-import bio.terra.tanagra.service.underlay.AttributeMapping;
-import bio.terra.tanagra.service.underlay.AttributeMapping.LookupColumn;
-import bio.terra.tanagra.service.underlay.AttributeMapping.SimpleColumn;
-import bio.terra.tanagra.service.underlay.Column;
-import bio.terra.tanagra.service.underlay.ForeignKey;
-import bio.terra.tanagra.service.underlay.Table;
-import bio.terra.tanagra.service.underlay.Underlay;
+import bio.terra.tanagra.underlay.AttributeMapping;
+import bio.terra.tanagra.underlay.AttributeMapping.LookupColumn;
+import bio.terra.tanagra.underlay.AttributeMapping.SimpleColumn;
+import bio.terra.tanagra.underlay.Column;
+import bio.terra.tanagra.underlay.ForeignKey;
+import bio.terra.tanagra.underlay.Table;
+import bio.terra.tanagra.underlay.Underlay;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableMap;
 import java.util.Map;
