+++ api/src/main/java/bio/terra/tanagra/underlay/BigQueryDataset.java
@@ -1,4 +1,4 @@
-package bio.terra.tanagra.service.underlay;
+package bio.terra.tanagra.underlay;
 
 import com.google.auto.value.AutoValue;
