+++ api/src/test/java/bio/terra/tanagra/service/underlay/NauticalUnderlayUtils.java
@@ -4,6 +4,11 @@ import bio.terra.tanagra.model.Attribute;
 import bio.terra.tanagra.model.DataType;
 import bio.terra.tanagra.model.Entity;
 import bio.terra.tanagra.model.Relationship;
+import bio.terra.tanagra.underlay.BigQueryDataset;
+import bio.terra.tanagra.underlay.Column;
+import bio.terra.tanagra.underlay.Table;
+import bio.terra.tanagra.underlay.Underlay;
+import bio.terra.tanagra.underlay.UnderlayConversion;
 import com.google.common.io.Resources;
 import com.google.protobuf.TextFormat;
 import java.io.IOException;
