+++ api/src/test/java/bio/terra/tanagra/service/underlay/UnderlayConversionTest.java
@@ -37,7 +37,12 @@ import bio.terra.tanagra.proto.underlay.EnumHint;
 import bio.terra.tanagra.proto.underlay.EnumHintValue;
 import bio.terra.tanagra.proto.underlay.FilterableAttribute;
 import bio.terra.tanagra.proto.underlay.IntegerBoundsHint;
-import bio.terra.tanagra.service.underlay.AttributeMapping.LookupColumn;
+import bio.terra.tanagra.underlay.AttributeMapping;
+import bio.terra.tanagra.underlay.AttributeMapping.LookupColumn;
+import bio.terra.tanagra.underlay.EntityFiltersSchema;
+import bio.terra.tanagra.underlay.ForeignKey;
+import bio.terra.tanagra.underlay.Underlay;
+import bio.terra.tanagra.underlay.UnderlayConversion;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableTable;
