+++ api/src/test/java/bio/terra/tanagra/service/underlay/UnderlayTest.java
@@ -10,6 +10,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import bio.terra.tanagra.model.Entity;
+import bio.terra.tanagra.underlay.Underlay;
 import java.util.Optional;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Tag;
